Add support for gym-mupen64plus
The Dockerfile needs to be updated to be based off of a gym-mupen64plus
image to use the dependencies for running a MarioKart64 environment. To
aid installation and usage, the README needs to be updated with the
relevant new information and a few default parameters need to be changed
to ensure the emulator can run smoothly.

Signed-Off-By: Robert Clark <robdclark@outlook.com>